+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [ita-challenges-frontend-3.7.0-RELEASE] - 2025-10-06
+
+### Added
+- Using the edit service Mentor can edit a current challenge by clicking on edit button (Taiga [#658], PR [#691]).
+- 
 ### [ita-challenges-frontend-3.6.0-RELEASE] - 2025-09-23
 
 ### Added